writes to file, however fails to read vec when more tha 1 value

+++ src/lib.rs
@@ -4,7 +4,7 @@ use std::fs;
 use std::fmt::Formatter;
 use std::fs::{OpenOptions, File};
 use std::path::Path;
-use std::io::{Write, BufReader};
+use std::io::{Write, BufReader, BufWriter, Read};
 use serde_json;
 use serde::{Serialize,Deserialize};
 use std::env::{set_current_dir, join_paths};
@@ -53,18 +53,16 @@ pub enum CmdType {
     Get,
     Rm,
 }
-
-pub fn read_kvscommands_from_file<P: AsRef<Path>>(path: P) -> Result<Vec<KvsCommand>> {
-    // Open the file in read-only mode with buffer.
-    let file = File::open(path).unwrap();
-    let reader = BufReader::new(file);
-
-    // Read the JSON contents of the file as an instance of `kvscmd`.
-    let kvscmd = serde_json::from_reader(reader).unwrap();
-
-    // Return the `kvscmd`.
-    Ok(kvscmd)
-}
+//
+// pub fn read_kvscommands_from_file<P: AsRef<Path>>(path: P) -> Result<Vec<KvsCommand>> {
+//     // Open the file in read-only mode with buffer.
+//
+//
+//     // Read the JSON contents of the file as an instance of `kvscmd`.
+//
+//     // Return the `kvscmd`.
+//     Ok(kvscmd)
+// }
 
 impl KvStore {
     pub fn new() -> KvStore{
@@ -74,17 +72,31 @@ impl KvStore {
         }
     }
     pub fn new_with_path(path: &path::Path) -> KvStore{
+        println!("create new Kvstore");
         KvStore{
             kv_db: HashMap::new(),
             path: path.to_str().unwrap().parse().unwrap()
         }
     }
     pub fn with_old_path(path: &path::Path) -> KvStore{
-        let deserialised = read_kvscommands_from_file(path).unwrap();
-       // let deserialised = serde_json::from_slice::<Vec<KvsCommand>>(&x).unwrap();
+        println!("read commands with old path");
+        let file = OpenOptions::new()
+            .read(true)
+            .append(true)
+            .open(&path)
+            .expect("Unable to open file");
+
+        let reader = BufReader::new(&file);
+
+        let mut deserialised: Vec<KvsCommand> =  serde_json::from_reader(reader).unwrap();
+
+       // let deserialised = read_kvscommands_from_file(path).unwrap();
+
         let mut kv_db:HashMap<String,String> = HashMap::new();
         for x in deserialised {
+            println!("loop through commands");
             if x.command== CmdType::Set{
+                println!("set command");
                 kv_db.insert(x.key,x.value.unwrap());
             }
         }
@@ -103,8 +115,16 @@ impl KvStore {
 
             let the_path = Path::new(&path).join("dbcmds.txt");
              println!("The full path on exists line  {:?}", the_path);
-             let  db_file = fs::File::create(the_path.clone()).unwrap();
-            return Result::Ok(KvStore::with_old_path(the_path.as_path()));
+            if the_path.is_file(){
+                println!("file exists, using it");
+
+                return Result::Ok(KvStore::with_old_path(the_path.as_path()));
+            }
+            else {
+                println!("file doesnt exist, making it so");
+                let db_file = fs::File::create(the_path.clone()).unwrap();
+                return Result::Ok(KvStore::new_with_path(the_path.as_path()));
+            }
         }
         fs::create_dir_all(&path);
         let the_path = Path::new(&path).join("dbcmds.txt");
@@ -121,11 +141,39 @@ impl KvStore {
             key,
             value: Option::from(value),
         };
-        let formatted = serde_json::to_string(&set_command).unwrap();
 
-        let mut file = fs::File::open(&self.path).unwrap();
+        let mut file = OpenOptions::new()
+            .read(true)
+            .append(true)
+            .open(&self.path)
+            .expect("Unable to open file");
+        let mut buffy = String::new();
+       file.read_to_string(&mut buffy);
 
-        file.write(formatted.as_ref());
+        let reader = BufReader::new(&file);
+        // Read the JSON contents of the file as an instance of `kvscmd`
+        if file.metadata().unwrap().len() <=0
+        {
+            println!("failed to unwrap, creating new vector");
+            let mut x = Vec::new(); x.push(set_command);
+            let mut file = BufWriter::new(&file);
+            let formatted_vec = serde_json::to_string(&x).unwrap();
+            println!("formatted vec: {}", formatted_vec);
+            file.write_all(formatted_vec.as_bytes()).expect("Unable to write data");
+            return Ok(true);
+        }
+
+        let mut kvscmd: Vec<KvsCommand> =  serde_json::from_str(&buffy).unwrap();
+
+        kvscmd.push(set_command);
+
+        let mut file = BufWriter::new(&file);
+
+        let formatted_vec = serde_json::to_string(&kvscmd).unwrap();
+        println!("formatted vec to write: {}", &formatted_vec);
+
+        file.write_all(formatted_vec.as_bytes()).expect("Unable to write data");
+
         Ok(true)
     }
